% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{snippet_themes}
\alias{snippet_themes}
\title{Return paths to themes included in the snippet package}
\usage{
snippet_themes()
}
\value{
A character vector of file paths to the themes
}
\description{
Return paths to themes included in the snippet package
}
\examples{
snippet_themes()
}
